chore: refactor && fixed tests
