+++ lib/domain/home/home_screen.dart
@@ -1,34 +1,249 @@
-import 'package:biely_kvet/domain/answers/answer_screen.dart';
-import 'package:biely_kvet/domain/core/adaptive/adaptive_tabbed_page.dart';
-import 'package:biely_kvet/domain/rules/rules_screen.dart';
+import 'dart:async';
+import 'dart:io';
+
+import 'package:adaptive_widgets_flutter/adaptive_widgets.dart';
+import 'package:biely_kvet/domain/home/rules/rules_screen.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+import 'answers/answer_screen.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return AdaptiveTabbedPage(
+    return TabbedPage(
       tabBarItems: [
         TabBarItem(
           // TODO: asi custom ikonky
           icon: const Icon(CupertinoIcons.question),
           label: 'Odpovede',
-          page: AdaptiveTabBarContent(pageContent: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: AnswerScreen(),
-          )),
+          page: TabBarContent(
+            pageContent: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: AnswerScreen(),
+            ),
+          ),
         ),
         TabBarItem(
           // TODO: asi custom ikonky
           icon: const Icon(CupertinoIcons.book),
           label: 'Pravidlá',
-          page: AdaptiveTabBarContent(pageContent: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: RulesScreen(),
-          )),
+          page: TabBarContent(
+            pageContent: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: RulesScreen(),
+            ),
+          ),
         ),
       ],
     );
+  }
+}
+
+class TabBarItem {
+  final Widget icon;
+  final Widget? activeIcon;
+  final String label;
+  final TabBarContent page;
+
+  TabBarItem({
+    required this.icon,
+    required this.label,
+    this.activeIcon,
+    required this.page,
+  });
+}
+
+class TabBarContent extends StatelessWidget {
+  final Widget pageContent;
+
+  const TabBarContent({super.key, required this.pageContent});
+
+  @override
+  Widget build(BuildContext context) {
+    return pageContent;
+  }
+}
+
+class TabbedPage extends StatefulWidget {
+  final List<TabBarItem> tabBarItems;
+
+  const TabbedPage({super.key, required this.tabBarItems});
+
+  @override
+  State<TabbedPage> createState() => _AdaptiveTabbedPageState();
+}
+
+class _AdaptiveTabbedPageState extends State<TabbedPage> {
+  int currentPageIndex = 0;
+
+  // TODO: skusit oddelit celu implementaciu timera do separatnej triedy
+  Timer? _timer;
+  int _remainingSeconds = 0;
+  int _initialSeconds = 0;
+  bool _isTimerActive = false;
+  static const _timerDuration = 60;
+
+  late StreamController<int> _events;
+
+  @override
+  initState() {
+    super.initState();
+    _events = StreamController<int>.broadcast();
+  }
+
+  @override
+  void dispose() {
+    _timer?.cancel();
+    _events.close();
+    super.dispose();
+  }
+
+  void _resetTimer() {
+    _timer?.cancel();
+    setState(() {
+      _remainingSeconds = _initialSeconds;
+    });
+  }
+
+  void _setSeconds(int value) {
+    if (value <= 0) return;
+
+    setState(() {
+      _initialSeconds = value;
+      _remainingSeconds = value;
+    });
+  }
+
+  void _start() {
+    if (_remainingSeconds <= 0) {
+      _setSeconds(_timerDuration);
+      if (_remainingSeconds <= 0) return;
+    }
+
+    _timer?.cancel();
+    _timer = Timer.periodic(const Duration(seconds: 1), (_) async {
+      if (!mounted) {
+        _timer?.cancel();
+        return;
+      }
+
+      if (_remainingSeconds <= 1) {
+        setState(() {
+          _remainingSeconds = 0;
+        });
+        _timer?.cancel();
+        await _showFinishedDialog();
+      } else {
+        setState(() {
+          _remainingSeconds--;
+          _events.add(_remainingSeconds);
+        });
+      }
+    });
+  }
+
+  Future<void> _showFinishedDialog() {
+    if (_isTimerActive) {
+      Navigator.of(context).pop();
+    }
+    return AdaptiveWidgets.showDialog(
+      context,
+      title: 'Časovač',
+      content: 'Koniec, bazerant.',
+      actionButtons: [
+        AdaptiveDialogButtonBuilder(
+          text: 'OK',
+          onPressed: (ctx) => Navigator.of(ctx).pop(),
+        ),
+      ],
+    );
+  }
+
+  String get _formatted {
+    final m = _remainingSeconds ~/ 60;
+    final s = _remainingSeconds % 60;
+    return '${m.toString().padLeft(2, '0')}:${s.toString().padLeft(2, '0')}';
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    var allPages = widget.tabBarItems.map((e) => e.page).toList();
+    return Scaffold(
+      floatingActionButton: FloatingActionButton(
+        onPressed: () async {
+          setState(() {
+            _isTimerActive = true;
+          });
+          _start();
+          await showDialog(
+            barrierDismissible: false,
+            context: context,
+            builder: (context) {
+              return StatefulBuilder(
+                builder: (context, setState) {
+                  var stopTimerLabel = const Text('Stop');
+                  return AlertDialog.adaptive(
+                        actions: [
+                          // TODO: asi pre iIOS toto nebude vhodne, pouzit CupertinoDialogAction
+                          Platform.isIOS
+                              ? CupertinoDialogAction(
+                                  onPressed: () {
+                                    _closeTimerDialog(setState, context);
+                                  },
+                                  child: stopTimerLabel,
+                                )
+                              : TextButton(
+                                  onPressed: () {
+                                    _closeTimerDialog(setState, context);
+                                  },
+                                  child: stopTimerLabel,
+                                ),
+                        ],
+                        title: const Text('Časovač'),
+                        content: StreamBuilder(
+                          stream: _events.stream,
+                          builder: (context, snapshot) => Text(_formatted),
+                        ),
+                      )
+                      as Widget;
+                },
+              );
+            },
+          );
+        },
+        shape: CircleBorder(),
+        child: Icon(Icons.timer),
+      ),
+      bottomNavigationBar: NavigationBar(
+        onDestinationSelected: (int index) {
+          setState(() {
+            currentPageIndex = index;
+          });
+        },
+        indicatorColor: Colors.amber,
+        selectedIndex: currentPageIndex,
+        destinations: widget.tabBarItems
+            .map(
+              (item) => NavigationDestination(
+                icon: item.icon,
+                label: item.label,
+                selectedIcon: item.activeIcon,
+              ),
+            )
+            .toList(),
+      ),
+      body: allPages[currentPageIndex],
+    );
+  }
+
+  void _closeTimerDialog(StateSetter setState, BuildContext context) {
+    setState(() {
+      _isTimerActive = false;
+    });
+    Navigator.of(context).pop();
+    _resetTimer();
   }
 }
