+++ test/app_test.dart
@@ -5,9 +5,7 @@ void main() {
   testWidgets('App načíta hlavné menu', (tester) async {
     await tester.pumpWidget(const BielyKvetApp());
 
-    expect(find.text('Biely kvet'), findsOneWidget);
-    expect(find.text('Hľadať odpoveď'), findsOneWidget);
-    expect(find.text('Časovač'), findsOneWidget);
-    expect(find.text('Pravidlá hry'), findsOneWidget);
+    expect(find.text('Odpovede'), findsOneWidget);
+    expect(find.text('Pravidlá'), findsOneWidget);
   });
 }
